Refactor `formatChainName`
It titleizes anything non-special thrown at it. For special cases, like BSC or XDAI we have custom nice names.
