+++ src/components/SalePointSelector/SalePointOption.tsx
@@ -8,7 +8,7 @@ import { COLORS } from '~/shared/utils/styled'
 import { Tick as PrestyledTick } from '~/shared/components/Checkbox'
 import NetworkIcon from '~/shared/components/NetworkIcon'
 import { getConfigForChain } from '~/shared/web3/config'
-import { formatChainName } from '~/shared/utils/chains'
+import { formatChainName } from '~/utils'
 import { useWalletAccount } from '~/shared/stores/wallet'
 import Select from '~/marketplace/components/SelectField2'
 import { getDataUnion, getDataUnionsOwnedByInChain } from '~/getters/du'
