+++ src/marketplace/containers/ProjectPage/Description.tsx
@@ -7,7 +7,7 @@ import Button from '~/shared/components/Button'
 import { getProjectTypeName } from '~/getters'
 import { ProjectType } from '~/shared/types'
 import FormattedPaymentRate from '~/components/FormattedPaymentRate'
-import { formatChainName } from '~/shared/utils/chains'
+import { formatChainName } from '~/utils'
 import { WhiteBox } from '~/shared/components/WhiteBox'
 import { getConfigForChain } from '~/shared/web3/config'
 import { timeUnits } from '~/shared/utils/timeUnit'
