+++ src/modals/AccessPeriodModal.tsx
@@ -8,7 +8,7 @@ import { COLORS, LIGHT, MEDIUM } from '~/shared/utils/styled'
 import { getConfigForChain } from '~/shared/web3/config'
 import NetworkIcon from '~/shared/components/NetworkIcon'
 import TokenLogo from '~/shared/components/TokenLogo'
-import { formatChainName } from '~/shared/utils/chains'
+import { formatChainName } from '~/utils'
 import { fromDecimals } from '~/marketplace/utils/math'
 import { priceForTimeUnits } from '~/marketplace/utils/price'
 import { address0 } from '~/consts'
