+++ src/modals/ChainSelectorModal.tsx
@@ -5,7 +5,7 @@ import Button from '~/shared/components/Button'
 import NetworkIcon from '~/shared/components/NetworkIcon'
 import { MEDIUM } from '~/shared/utils/styled'
 import { getConfigForChain } from '~/shared/web3/config'
-import { formatChainName } from '~/shared/utils/chains'
+import { formatChainName } from '~/utils'
 import useIsMounted from '~/shared/hooks/useIsMounted'
 import { getCustomTokenBalance } from '~/marketplace/utils/web3'
 import { getTokenInfo } from '~/hooks/useTokenInfo'
