+++ src/pages/ProjectPage/AccessManifest.tsx
@@ -5,7 +5,7 @@ import { REGULAR, TABLET } from '~/shared/utils/styled'
 import Button from '~/shared/components/Button'
 import { ProjectType, SalePoint } from '~/shared/types'
 import FormattedPaymentRate from '~/components/FormattedPaymentRate'
-import { formatChainName } from '~/shared/utils/chains'
+import { formatChainName } from '~/utils'
 import { WhiteBox } from '~/shared/components/WhiteBox'
 import { getConfigForChain } from '~/shared/web3/config'
 import { timeUnits } from '~/shared/utils/timeUnit'
